fix(acr): enforce numeric and invocation contracts
24号计划 §5：提交前统一校验 buffer_count/scalar_bytes/domain/backend launcher/NumericPolicy（validate_invocation）；ScalarArgBlob 改 memcpy 对齐读取（read_scalar，消除 reinterpret_cast 未对齐 UB）；Reduction 声明 FP64 accumulator 与 CPU(double 局部累加)/CUDA(double partials+atomicAdd) 真实实现一致，桥接 REDUCE partials 改 double；新增契约/对齐/FP64 归约测试，51 项相关测试全通过。

+++ lib/acr/backends/cuda/bridge/acr_cuda_bridge.h
@@ -61,13 +61,13 @@ ACR_CUDA_BRIDGE_API int acr_cuda_executor_submit_copy(
     float* y, const float* x,
     uint64_t* elapsed_ns, const char** last_error);
 
-// 归约：partials[chunk_index * blocks + blockIdx] = sum(x[begin..end) 的 block 和)
+// 归约（FP64 累加）：partials[chunk_index * blocks + blockIdx] = sum(x[begin..end) 的 block 和)
 // blocks_per_chunk 固定（如 256），merge 阶段累加全部 partials。
 ACR_CUDA_BRIDGE_API int acr_cuda_executor_submit_reduce(
     void* handle,
     size_t begin, size_t end,
     const float* x,
-    float* partials,
+    double* partials,
     size_t blocks_per_chunk,
     uint64_t chunk_index,
     uint64_t* elapsed_ns, const char** last_error);
